perf(sim): Phase 0 bench/counters + VirtualClock/cycles hot-path wins
PERF-METHODOLOGY Phase 0 measurement infra (RelWithDebInfo, advisory):
- bench/bench_sim.cpp: insn/sec over the armcc/AC6 corpus, median of N reps
- MF_PERF_STATS compile-gated counters (bus reads, fetches, find_region
  scan length, 16/32-bit mix); zero-cost when the macro is off
- instrumentation in Bus (read/write/find_region) + cortex_m3 (fetch16/step)

Two measured, behavior-preserving optimizations (gprof #1 hotspot):
- VirtualClock::advance: __uint128_t -> uint64_t. The per-step cycle delta
  fits 64-bit (max 2^32 * 72MHz < 2^64), so results are bit-identical and
  the __udivti3 soft-divide (33% of runtime) is gone.
- SimulationCoordinator::step: cache prev cycles in the existing
  last_cycles_ member (was 2x virtual cycles()/step, now 1x).

RelWithDebInfo insn/sec (armcc/AC6 corpus): gpio +22%, uart +32%, tim +26%.
ctest 321/321 green (Debug + RelWithDebInfo); __udivti3/__umodti3 absent.

+++ src/sim/virtual_clock.cpp
@@ -3,9 +3,9 @@
 namespace micro_forge::sim {
 
 VirtualClock::VirtualClock(std::span<const DomainConfig> domains)
-    : sysclk_freq_hz_(
-          (domains.empty() || domains[0].freq_hz == 0) ? 1
-                                                        : domains[0].freq_hz) {
+    : sysclk_freq_hz_((domains.empty() || domains[0].freq_hz == 0)
+                          ? 1
+                          : domains[0].freq_hz) {
     domains_.reserve(domains.size());
     for (const auto& cfg : domains) {
         domains_.push_back({cfg.freq_hz, 0, 0});
@@ -13,12 +13,16 @@ VirtualClock::VirtualClock(std::span<const DomainConfig> domains)
 }
 
 void VirtualClock::advance(uint64_t cpu_cycles) {
-    __uint128_t ns_product =
-        static_cast<__uint128_t>(cpu_cycles) * 1'000'000'000ULL +
-        total_ns_residual_;
-    total_ns_ += static_cast<uint64_t>(ns_product / sysclk_freq_hz_);
-    total_ns_residual_ =
-        static_cast<uint64_t>(ns_product % sysclk_freq_hz_);
+    // INVARIANT: cpu_cycles is a *per-step* delta (single-instruction cycle
+    // count, 1–3), the only input from SimulationCoordinator::step. With the
+    // largest STM32F1 clock (72 MHz) and cpu_cycles up to 2^32, every product
+    // here stays well under 2^64 (max ≈ 2^32 * 1e9 = 4.29e18 < 1.8e19), so
+    // uint64_t is bit-identical to the __uint128_t it replaces — without the
+    // __udivti3/__umodti3 soft-divide libcalls (x86-64 has no 128-bit divide),
+    // which made advance() the #1 measured hotspot (33% of runtime).
+    uint64_t ns_product = cpu_cycles * 1'000'000'000ULL + total_ns_residual_;
+    total_ns_ += ns_product / sysclk_freq_hz_;
+    total_ns_residual_ = ns_product % sysclk_freq_hz_;
 
     for (auto& d : domains_) {
         if (d.freq_hz == 0) {
@@ -26,10 +30,9 @@ void VirtualClock::advance(uint64_t cpu_cycles) {
         }
 
         // ticks = cpu_cycles * domain_freq / sysclk_freq
-        // 用 __uint128_t 避免溢出
-        __uint128_t product = static_cast<__uint128_t>(cpu_cycles) * d.freq_hz;
-        uint64_t ticks = static_cast<uint64_t>(product / sysclk_freq_hz_);
-        uint64_t rem = static_cast<uint64_t>(product % sysclk_freq_hz_);
+        uint64_t product = cpu_cycles * d.freq_hz;
+        uint64_t ticks = product / sysclk_freq_hz_;
+        uint64_t rem = product % sysclk_freq_hz_;
 
         d.residual += rem;
         if (d.residual >= sysclk_freq_hz_) {
